Feat/task rbac (#27)
* feat: add error handling and user role validation in task service

* pre-commit

* feat: unit tests and changes found by them

* pre-commit

* changes to pr

* pre-commit

* changes to pr

* Correctly apply pagination and sorting using utils function

* Refactor API documentation for group and task routes to improve clarity and consistency

* pre-commit

* pre-commit

* feat: sanitize branch name for Docker tag in workflow

* fixes

* fix docs

---------

+++ package/service/group_service.go
@@ -5,17 +5,21 @@ import (
 
 	"github.com/mini-maxit/backend/package/domain/models"
 	"github.com/mini-maxit/backend/package/domain/schemas"
+	"github.com/mini-maxit/backend/package/domain/types"
+	"github.com/mini-maxit/backend/package/errors"
 	"github.com/mini-maxit/backend/package/repository"
 	"github.com/mini-maxit/backend/package/utils"
 	"gorm.io/gorm"
 )
 
 type GroupService interface {
-	CreateGroup(tx *gorm.DB, group *schemas.Group) (int64, error)
-	DeleteGroup(tx *gorm.DB, groupId int64) error
-	Edit(tx *gorm.DB, groupId int64, editInfo *schemas.EditGroup) (*schemas.Group, error)
-	GetAllGroup(tx *gorm.DB, queryParams map[string]interface{}) ([]schemas.Group, error)
-	GetGroup(tx *gorm.DB, groupId int64) (*schemas.Group, error)
+	CreateGroup(tx *gorm.DB, current_user schemas.User, group *schemas.Group) (int64, error)
+	DeleteGroup(tx *gorm.DB, current_user schemas.User, groupId int64) error
+	Edit(tx *gorm.DB, current_user schemas.User, groupId int64, editInfo *schemas.EditGroup) (*schemas.Group, error)
+	GetAllGroup(tx *gorm.DB, current_user schemas.User, queryParams map[string]interface{}) ([]schemas.Group, error)
+	GetGroup(tx *gorm.DB, current_user schemas.User, groupId int64) (*schemas.Group, error)
+	AddUsersToGroup(tx *gorm.DB, current_user schemas.User, groupId int64, userIds []int64) error
+	GetGroupUsers(tx *gorm.DB, current_user schemas.User, groupId int64) ([]schemas.User, error)
 }
 
 var (
@@ -26,9 +30,16 @@ var (
 
 type groupService struct {
 	groupRepository repository.GroupRepository
+	userRepository  repository.UserRepository
+	userService     UserService
 }
 
-func (gs *groupService) CreateGroup(tx *gorm.DB, group *schemas.Group) (int64, error) {
+func (gs *groupService) CreateGroup(tx *gorm.DB, current_user schemas.User, group *schemas.Group) (int64, error) {
+	err := utils.ValidateRoleAccess(current_user.Role, []types.UserRole{types.UserRoleAdmin, types.UserRoleTeacher})
+	if err != nil {
+		return 0, err
+	}
+
 	validate, err := utils.NewValidator()
 	if err != nil {
 		return -1, err
@@ -50,11 +61,30 @@ func (gs *groupService) CreateGroup(tx *gorm.DB, group *schemas.Group) (int64, e
 	return groupId, nil
 }
 
-func (gs *groupService) DeleteGroup(tx *gorm.DB, groupId int64) error {
+func (gs *groupService) DeleteGroup(tx *gorm.DB, current_user schemas.User, groupId int64) error {
+	err := utils.ValidateRoleAccess(current_user.Role, []types.UserRole{types.UserRoleAdmin, types.UserRoleTeacher})
+	if err != nil {
+		return err
+	}
+
+	group, err := gs.groupRepository.GetGroup(tx, groupId)
+	if err != nil {
+		return err
+	}
+
+	if current_user.Role == types.UserRoleTeacher && group.CreatedBy != current_user.Id {
+		return errors.ErrNotAuthorized
+	}
+
 	return gs.groupRepository.DeleteGroup(tx, groupId)
 }
 
-func (gs *groupService) Edit(tx *gorm.DB, groupId int64, editInfo *schemas.EditGroup) (*schemas.Group, error) {
+func (gs *groupService) Edit(tx *gorm.DB, current_user schemas.User, groupId int64, editInfo *schemas.EditGroup) (*schemas.Group, error) {
+	err := utils.ValidateRoleAccess(current_user.Role, []types.UserRole{types.UserRoleAdmin, types.UserRoleTeacher})
+	if err != nil {
+		return nil, err
+	}
+
 	validate, err := utils.NewValidator()
 	if err != nil {
 		return nil, err
@@ -70,6 +100,10 @@ func (gs *groupService) Edit(tx *gorm.DB, groupId int64, editInfo *schemas.EditG
 		}
 		return nil, err
 	}
+	if current_user.Role == types.UserRoleTeacher && model.CreatedBy != current_user.Id {
+		return nil, errors.ErrNotAuthorized
+	}
+
 	gs.updateModel(model, editInfo)
 
 	newModel, err := gs.groupRepository.Edit(tx, groupId, model)
@@ -77,18 +111,34 @@ func (gs *groupService) Edit(tx *gorm.DB, groupId int64, editInfo *schemas.EditG
 		return nil, err
 	}
 	return gs.modelToSchema(newModel), nil
-
 }
 
-func (gs *groupService) GetAllGroup(tx *gorm.DB, queryParams map[string]interface{}) ([]schemas.Group, error) {
-	limit := queryParams["limit"].(uint64)
-	offset := queryParams["offset"].(uint64)
-	sort := queryParams["sort"].(string)
-	groups, err := gs.groupRepository.GetAllGroup(tx, int(offset), int(limit), sort)
+func (gs *groupService) GetAllGroup(tx *gorm.DB, current_user schemas.User, queryParams map[string]interface{}) ([]schemas.Group, error) {
+	err := utils.ValidateRoleAccess(current_user.Role, []types.UserRole{types.UserRoleAdmin, types.UserRoleTeacher})
 	if err != nil {
 		return nil, err
 	}
 
+	limit := queryParams["limit"].(uint64)
+	offset := queryParams["offset"].(uint64)
+	sort := queryParams["sort"].(string)
+	if sort == "" {
+		sort = "created_at:desc"
+	}
+	var groups []models.Group
+
+	switch current_user.Role {
+	case types.UserRoleAdmin:
+		groups, err = gs.groupRepository.GetAllGroup(tx, int(offset), int(limit), sort)
+		if err != nil {
+			return nil, err
+		}
+	case types.UserRoleTeacher:
+		groups, err = gs.groupRepository.GetAllGroupForTeacher(tx, current_user.Id, int(offset), int(limit), sort)
+		if err != nil {
+			return nil, err
+		}
+	}
 	var result []schemas.Group
 	for _, group := range groups {
 		result = append(result, *gs.modelToSchema(&group))
@@ -97,7 +147,12 @@ func (gs *groupService) GetAllGroup(tx *gorm.DB, queryParams map[string]interfac
 	return result, nil
 }
 
-func (gs *groupService) GetGroup(tx *gorm.DB, groupId int64) (*schemas.Group, error) {
+func (gs *groupService) GetGroup(tx *gorm.DB, current_user schemas.User, groupId int64) (*schemas.Group, error) {
+	err := utils.ValidateRoleAccess(current_user.Role, []types.UserRole{types.UserRoleAdmin, types.UserRoleTeacher})
+	if err != nil {
+		return nil, err
+	}
+
 	group, err := gs.groupRepository.GetGroup(tx, groupId)
 	if err != nil {
 		if err == gorm.ErrRecordNotFound {
@@ -106,7 +161,69 @@ func (gs *groupService) GetGroup(tx *gorm.DB, groupId int64) (*schemas.Group, er
 		return nil, err
 	}
 
+	if current_user.Role == types.UserRoleTeacher && group.CreatedBy != current_user.Id {
+		return nil, errors.ErrNotAuthorized
+	}
+
 	return gs.modelToSchema(group), nil
+}
+
+func (gs *groupService) AddUsersToGroup(tx *gorm.DB, current_user schemas.User, groupId int64, userIds []int64) error {
+	err := utils.ValidateRoleAccess(current_user.Role, []types.UserRole{types.UserRoleAdmin, types.UserRoleTeacher})
+	if err != nil {
+		return err
+	}
+
+	group, err := gs.groupRepository.GetGroup(tx, groupId)
+	if err != nil {
+		return err
+	}
+
+	if current_user.Role == types.UserRoleTeacher && group.CreatedBy != current_user.Id {
+		return errors.ErrNotAuthorized
+	}
+
+	for _, userId := range userIds {
+		_, err := gs.userRepository.GetUser(tx, userId)
+		if err != nil {
+			return err
+		}
+
+		err = gs.groupRepository.AddUserToGroup(tx, groupId, userId)
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func (gs *groupService) GetGroupUsers(tx *gorm.DB, current_user schemas.User, groupId int64) ([]schemas.User, error) {
+	err := utils.ValidateRoleAccess(current_user.Role, []types.UserRole{types.UserRoleAdmin, types.UserRoleTeacher})
+	if err != nil {
+		return nil, err
+	}
+
+	group, err := gs.groupRepository.GetGroup(tx, groupId)
+	if err != nil {
+		return nil, err
+	}
+
+	if current_user.Role == types.UserRoleTeacher && group.CreatedBy != current_user.Id {
+		return nil, errors.ErrNotAuthorized
+	}
+
+	users, err := gs.groupRepository.GetGroupUsers(tx, groupId)
+	if err != nil {
+		return nil, err
+	}
+
+	var result []schemas.User
+	for _, user := range users {
+		result = append(result, *gs.userService.modelToSchema(&user))
+	}
+
+	return result, nil
 }
 
 func (gs *groupService) updateModel(model *models.Group, editInfo *schemas.EditGroup) {
@@ -125,8 +242,10 @@ func (gs *groupService) modelToSchema(model *models.Group) *schemas.Group {
 	}
 }
 
-func NewGroupService(groupRepository repository.GroupRepository) GroupService {
+func NewGroupService(groupRepository repository.GroupRepository, userRepository repository.UserRepository, userService UserService) GroupService {
 	return &groupService{
 		groupRepository: groupRepository,
+		userRepository:  userRepository,
+		userService:     userService,
 	}
 }
